+++ CHANGELOG.md
@@ -59,6 +59,8 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 * (rpc) [tharsis#679](https://github.com/tharsis/ethermint/pull/679) Fix file close handle.
 * (rpc) [tharsis#671](https://github.com/tharsis/ethermint/pull/671) Don't pass base fee externally for `EthCall`/`EthEstimateGas` apis.
+* (evm) [tharsis#674](https://github.com/tharsis/ethermint/pull/674) Refactor `ApplyMessage`, remove
+  `ApplyNativeMessage`.
 
 ## [v0.7.1] - 2021-10-08